Move main process
